feat(Store): add Store#receivePayload(payload: Payload)
We introduce a new API `receivePayload(payload): void`.
It provide write/update timing to Store.

Core concept of almin store is read-stack on CQRS.

Read-stack is usually consist of write-in-read and read-in-read.

write-in-read receive tha event(payload) from Write-stack.
In the almin, UseCase(Write-stack) dispatch a payload.
When just time, Store receive the payload and update own state.
In other word, write stack update the view model in read stack.

read-in-read just return own state of the Store.
When View needed application states, StoreGroup call each Store#getState and merge it.
In other word, just return the view model in read stack.